Hide TreasureChestItemIterator as an private inner class of TreasureChest. Clean-up 'findNextIdx' function.
