--- iterator/src/main/java/com/iluwatar/iterator/TreasureChestItemIterator.java
@@ -1,78 +1,0 @@
-/**
- * The MIT License
- * Copyright (c) 2014-2016 Ilkka Seppälä
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package com.iluwatar.iterator;
-
-import java.util.List;
-
-/**
- * 
- * TreasureChestItemIterator
- *
- */
-public class TreasureChestItemIterator implements ItemIterator {
-
-  private TreasureChest chest;
-  private int idx;
-  private ItemType type;
-
-  /**
-   * Constructor
-   */
-  public TreasureChestItemIterator(TreasureChest chest, ItemType type) {
-    this.chest = chest;
-    this.type = type;
-    this.idx = -1;
-  }
-
-  @Override
-  public boolean hasNext() {
-    return findNextIdx() != -1;
-  }
-
-  @Override
-  public Item next() {
-    idx = findNextIdx();
-    if (idx != -1) {
-      return chest.getItems().get(idx);
-    }
-    return null;
-  }
-
-  private int findNextIdx() {
-
-    List<Item> items = chest.getItems();
-    boolean found = false;
-    int tempIdx = idx;
-    while (!found) {
-      tempIdx++;
-      if (tempIdx >= items.size()) {
-        tempIdx = -1;
-        break;
-      }
-      if (type.equals(ItemType.ANY) || items.get(tempIdx).getType().equals(type)) {
-        break;
-      }
-    }
-    return tempIdx;
-  }
-}
